+++ contracts/test-utils/src/number_utils.ts
@@ -40,10 +40,7 @@ export function getRandomPortion(total: Numberish): BigNumber {
 export function getRandomFloat(min: Numberish, max: Numberish): BigNumber {
     // Generate a really high precision number between [0, 1]
     const r = new BigNumber(crypto.randomBytes(32).toString('hex'), 16).dividedBy(new BigNumber(2).pow(256).minus(1));
-    return new BigNumber(max)
-        .minus(min)
-        .times(r)
-        .plus(min);
+    return new BigNumber(max).minus(min).times(r).plus(min);
 }
 
 export const FIXED_POINT_BASE = new BigNumber(2).pow(127);
@@ -66,7 +63,7 @@ export function fromFixed(n: Numberish): BigNumber {
  * Converts two decimal numbers to integers with `precision` digits, then returns
  * the absolute difference.
  */
-export function getNumericalDivergence(a: Numberish, b: Numberish, precision: number = 18): number {
+export function getNumericalDivergence(a: Numberish, b: Numberish, precision = 18): number {
     const _a = new BigNumber(a);
     const _b = new BigNumber(b);
     const maxIntegerDigits = Math.max(
@@ -77,16 +74,13 @@ export function getNumericalDivergence(a: Numberish, b: Numberish, precision: nu
         const base = 10 ** (precision - maxIntegerDigits);
         return n.times(base).integerValue(BigNumber.ROUND_DOWN);
     };
-    return _toInteger(_a)
-        .minus(_toInteger(_b))
-        .abs()
-        .toNumber();
+    return _toInteger(_a).minus(_toInteger(_b)).abs().toNumber();
 }
 
 /**
  * Asserts that two numbers are equal up to `precision` digits.
  */
-export function assertRoughlyEquals(actual: Numberish, expected: Numberish, precision: number = 18): void {
+export function assertRoughlyEquals(actual: Numberish, expected: Numberish, precision = 18): void {
     if (getNumericalDivergence(actual, expected, precision) <= 1) {
         return;
     }
@@ -96,16 +90,8 @@ export function assertRoughlyEquals(actual: Numberish, expected: Numberish, prec
 /**
  * Asserts that two numbers are equal with up to `maxError` difference between them.
  */
-export function assertIntegerRoughlyEquals(
-    actual: Numberish,
-    expected: Numberish,
-    maxError: number = 1,
-    msg?: string,
-): void {
-    const diff = new BigNumber(actual)
-        .minus(expected)
-        .abs()
-        .toNumber();
+export function assertIntegerRoughlyEquals(actual: Numberish, expected: Numberish, maxError = 1, msg?: string): void {
+    const diff = new BigNumber(actual).minus(expected).abs().toNumber();
     if (diff <= maxError) {
         return;
     }