+++ contracts/test-utils/test/mocha_blockchain.ts
@@ -4,6 +4,7 @@ import * as process from 'process';
 import { expect } from '../src/chai_setup';
 import { constants } from '../src/constants';
 import { blockchainTests, describe } from '../src/mocha_blockchain';
+import { append } from './subtests/mocha_blockchain_1';
 
 blockchainTests('mocha blockchain extensions', env => {
     describe('blockchainTests()', () => {
@@ -78,7 +79,7 @@ blockchainTests('mocha blockchain extensions', env => {
         });
 
         describe('subtests', () => {
-            require('./subtests/mocha_blockchain_1').append(env);
+            append(env);
         });
     });
 
@@ -96,7 +97,7 @@ blockchainTests('mocha blockchain extensions', env => {
 function createHookedObject(obj: any, handler: (name: string) => void, methods: string[]): any {
     const hookedMethods = _.map(methods, methodName => {
         // tslint:disable: only-arrow-functions
-        return function(this: any, ...args: any[]): any {
+        return function (this: any, ...args: any[]): any {
             handler(methodName);
             return obj[methodName].call(this, ...args);
         };