+++ contracts/erc20/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v3.3.44 - _September 21, 2022_
+
+    * Dependencies updated
+
+## v3.3.43 - _September 12, 2022_
+
+    * Dependencies updated
+
 ## v3.3.42 - _September 7, 2022_
 
     * Dependencies updated