+++ contracts/test-utils/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v5.4.35 - _September 21, 2022_
+
+    * Dependencies updated
+
+## v5.4.34 - _September 12, 2022_
+
+    * Dependencies updated
+
 ## v5.4.33 - _September 7, 2022_
 
     * Dependencies updated