+++ contracts/utils/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v4.8.25 - _September 21, 2022_
+
+    * Dependencies updated
+
+## v4.8.24 - _September 12, 2022_
+
+    * Dependencies updated
+
 ## v4.8.23 - _September 7, 2022_
 
     * Dependencies updated